requirements.txt
aiogram==3.13.1
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sentry-sdk==1.40.1
